Update angular to 4.3.0, add testing api service
