+++ demo/portal/package.json
@@ -13,21 +13,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.2.6",
+    "@angular/animations": "^4.3.0",
     "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
+    "@angular/common": "^4.3.0",
+    "@angular/compiler": "^4.3.0",
+    "@angular/core": "^4.3.0",
+    "@angular/forms": "^4.3.0",
+    "@angular/http": "^4.3.0",
     "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/platform-browser": "^4.3.0",
+    "@angular/platform-browser-dynamic": "^4.3.0",
+    "@angular/router": "^4.3.0",
+    "angular-in-memory-web-api": "^0.3.2",
     "@raincatcher/wfm": "1.0.0",
     "bootstrap": "^3.3.7",
     "core-js": "^2.4.1",
-    "hammerjs": "^2.0.8",
     "jquery": "^3.2.1",
     "ngx-bootstrap": "^1.7.1",
     "patternfly": "^3.25.1",
@@ -36,16 +36,11 @@
   },
   "devDependencies": {
     "@angular/cli": "1.2.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/language-service": "^4.0.0",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.0.1",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3",
-    "webpack-dev-server": "~2.4.5",
-    "webpack": "~2.4.0",
+    "@angular/compiler-cli": "^4.3.0",
+    "@angular/language-service": "^4.3.0",
+    "@types/node": "^6.0.84",
     "autoprefixer": "^6.5.3",
+    "codelyzer": "~3.0.1",
     "css-loader": "^0.28.1",
     "cssnano": "^3.10.0",
     "exports-loader": "^0.6.3",
@@ -58,9 +53,13 @@
     "sass-loader": "^6.0.3",
     "script-loader": "^0.7.0",
     "source-map-loader": "^0.2.0",
-    "istanbul-instrumenter-loader": "^2.0.0",
     "style-loader": "^0.13.1",
     "stylus-loader": "^3.0.1",
-    "url-loader": "^0.5.7"
+    "ts-node": "~3.0.4",
+    "tslint": "~5.3.2",
+    "typescript": "~2.3.3",
+    "url-loader": "^0.5.7",
+    "webpack": "~2.4.0",
+    "webpack-dev-server": "~2.4.5"
   }
 }
