+++ demo/portal/src/app/app.module.ts
@@ -1,8 +1,8 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { NgModule } from '@angular/core';
+import { NgModule, Provider } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-
+import { HttpClientModule } from '@angular/common/http';
 
 import { ButtonsModule } from 'ngx-bootstrap';
 
@@ -13,6 +13,12 @@ import { NavigationComponent } from './navigation/navigation.component';
 import { NavLinkComponent } from './nav-link/nav-link.component';
 import { ProcessComponent } from './process/process.component';
 import { FooterComponent } from './footer/footer.component';
+
+import { InMemoryDataService } from './data/in-memory-data.service';
+
+import { environment } from '../environments/environment';
+import { ConfigService } from './config.service';
+
 // Router and routes setup
 const routes: Routes = [
   { path: 'process-instances', component: ProcessInstanceComponent },
@@ -20,6 +26,13 @@ const routes: Routes = [
   { path: 'tasks', component: TaskComponent },
   { path: '', redirectTo: '/process-instances', pathMatch: 'full' }
 ]
+
+const providers: Provider[] = [
+  ConfigService
+];
+if (!environment.production) {
+  providers.push(InMemoryDataService);
+}
 
 @NgModule({
   declarations: [
@@ -34,13 +47,14 @@ const routes: Routes = [
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
+    HttpClientModule,
     ButtonsModule.forRoot(),
     RouterModule.forRoot(
       routes,
       { enableTracing: true }
     )
   ],
-  providers: [],
+  providers: providers,
   bootstrap: [AppComponent]
 })
 export class AppModule { }
