+++ demo/portal/src/app/config.service.ts
@@ -1,0 +1,7 @@
+import { environment } from '../environments/environment';
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class ConfigService {
+  public apiRootUrl = environment.apiUrl;
+}
